refactor controller and service and update specs
